Migrate orders, products & metaboxes class to new dir
